fix credential helpers for clip v2 (#1481)
