+++ pkg/registry/credentials_test.go
