--- pkg/registry/oci_credentials.go
@@ -1,417 +1,0 @@
-package registry
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"os"
-	"strings"
-
-	"github.com/beam-cloud/clip/pkg/common"
-	"github.com/google/go-containerregistry/pkg/authn"
-)
-
-// CredType represents the type of registry credentials
-type CredType string
-
-const (
-	CredTypePublic  CredType = "public"
-	CredTypeBasic   CredType = "basic"
-	CredTypeAWS     CredType = "aws"
-	CredTypeGCP     CredType = "gcp"
-	CredTypeAzure   CredType = "azure"
-	CredTypeToken   CredType = "token"
-	CredTypeUnknown CredType = "unknown"
-)
-
-// Known credential environment variable keys
-var knownCredKeys = map[string]bool{
-	"USERNAME":                       true,
-	"PASSWORD":                       true,
-	"AWS_ACCESS_KEY_ID":              true,
-	"AWS_SECRET_ACCESS_KEY":          true,
-	"AWS_SESSION_TOKEN":              true,
-	"AWS_REGION":                     true,
-	"GOOGLE_APPLICATION_CREDENTIALS": true,
-	"GCP_PROJECT_ID":                 true,
-	"GCP_ACCESS_TOKEN":               true,
-	"AZURE_CLIENT_ID":                true,
-	"AZURE_CLIENT_SECRET":            true,
-	"AZURE_TENANT_ID":                true,
-	"DOCKER_USERNAME":                true,
-	"DOCKER_PASSWORD":                true,
-	"REGISTRY_USERNAME":              true,
-	"REGISTRY_PASSWORD":              true,
-	"NGC_API_KEY":                    true,
-	"GITHUB_TOKEN":                   true,
-	"DOCKERHUB_TOKEN":                true,
-}
-
-// Token credential keys
-var tokenCredKeys = []string{"NGC_API_KEY", "GITHUB_TOKEN", "DOCKERHUB_TOKEN"}
-
-// ParseRegistry extracts the registry hostname from an image reference
-// Examples:
-//   - "docker.io/library/ubuntu:20.04" -> "docker.io"
-//   - "gcr.io/project/image:tag" -> "gcr.io"
-//   - "registry.example.com:5000/image:tag" -> "registry.example.com:5000"
-//   - "ubuntu:20.04" -> "docker.io" (default)
-func ParseRegistry(imageRef string) string {
-	if imageRef == "" {
-		return ""
-	}
-
-	// Remove any transport prefix (docker://, oci://, etc.)
-	imageRef = strings.TrimPrefix(imageRef, "docker://")
-	imageRef = strings.TrimPrefix(imageRef, "oci://")
-
-	// Split by / to separate registry from repository
-	parts := strings.Split(imageRef, "/")
-
-	if len(parts) == 1 {
-		// No slash means it's a Docker Hub official image (e.g., "ubuntu:20.04")
-		return "docker.io"
-	}
-
-	// Check if first part looks like a registry (contains . or : or is localhost)
-	firstPart := parts[0]
-	if strings.Contains(firstPart, ".") ||
-		strings.Contains(firstPart, ":") ||
-		firstPart == "localhost" {
-		return firstPart
-	}
-
-	// Otherwise it's a Docker Hub user image (e.g., "username/image")
-	return "docker.io"
-}
-
-// ParseCredentialsFromEnv filters a map to only include known credential keys
-func ParseCredentialsFromEnv(envMap map[string]string) map[string]string {
-	creds := make(map[string]string)
-	for key, value := range envMap {
-		if knownCredKeys[key] {
-			creds[key] = value
-		}
-	}
-	return creds
-}
-
-// DetectCredentialType determines the type of credentials based on the registry and credential keys
-func DetectCredentialType(registry string, creds map[string]string) CredType {
-	if len(creds) == 0 {
-		return CredTypePublic
-	}
-
-	// Check for AWS credentials
-	if _, hasAwsKey := creds["AWS_ACCESS_KEY_ID"]; hasAwsKey {
-		if _, hasAwsSecret := creds["AWS_SECRET_ACCESS_KEY"]; hasAwsSecret {
-			return CredTypeAWS
-		}
-	}
-
-	// Check for GCP credentials
-	if _, hasGcp := creds["GOOGLE_APPLICATION_CREDENTIALS"]; hasGcp {
-		return CredTypeGCP
-	}
-	if _, hasGcpProject := creds["GCP_PROJECT_ID"]; hasGcpProject {
-		return CredTypeGCP
-	}
-	if _, hasGcpToken := creds["GCP_ACCESS_TOKEN"]; hasGcpToken {
-		return CredTypeGCP
-	}
-
-	// Check for Azure credentials
-	if _, hasAzureClientId := creds["AZURE_CLIENT_ID"]; hasAzureClientId {
-		if _, hasAzureSecret := creds["AZURE_CLIENT_SECRET"]; hasAzureSecret {
-			return CredTypeAzure
-		}
-	}
-
-	// Check for token-based auth (before basic auth)
-	for _, key := range tokenCredKeys {
-		if _, hasToken := creds[key]; hasToken {
-			return CredTypeToken
-		}
-	}
-
-	// Check for basic auth (username/password)
-	hasUsername := false
-	hasPassword := false
-
-	for key := range creds {
-		keyUpper := strings.ToUpper(key)
-		if strings.Contains(keyUpper, "USERNAME") {
-			hasUsername = true
-		}
-		if strings.Contains(keyUpper, "PASSWORD") {
-			hasPassword = true
-		}
-	}
-
-	if hasUsername && hasPassword {
-		return CredTypeBasic
-	}
-
-	// Detect based on registry
-	registry = strings.ToLower(registry)
-	if strings.Contains(registry, "ecr") || strings.Contains(registry, "amazonaws.com") {
-		return CredTypeAWS
-	}
-	if strings.Contains(registry, "gcr.io") || strings.Contains(registry, "pkg.dev") {
-		return CredTypeGCP
-	}
-	if strings.Contains(registry, "azurecr.io") {
-		return CredTypeAzure
-	}
-
-	return CredTypeUnknown
-}
-
-// MarshalCredentials serializes credentials to JSON format
-// The JSON structure includes the registry and credential type for easier retrieval
-func MarshalCredentials(registry string, credType CredType, creds map[string]string) (string, error) {
-	data := map[string]interface{}{
-		"registry":    registry,
-		"type":        string(credType),
-		"credentials": creds,
-	}
-
-	jsonBytes, err := json.Marshal(data)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal credentials: %w", err)
-	}
-
-	return string(jsonBytes), nil
-}
-
-// CreateSecretName generates a consistent secret name for a registry
-// Format: "oci-registry-<normalized-registry-name>"
-func CreateSecretName(registry string) string {
-	// Normalize registry name for use in secret name
-	normalized := strings.ToLower(registry)
-	normalized = strings.ReplaceAll(normalized, ".", "-")
-	normalized = strings.ReplaceAll(normalized, ":", "-")
-	normalized = strings.ReplaceAll(normalized, "/", "-")
-
-	return fmt.Sprintf("oci-registry-%s", normalized)
-}
-
-// CreateProviderFromCredentials creates a CLIP-compatible credential provider from a credential map
-// This function creates an appropriate credential provider without setting environment variables
-// Returns common.RegistryCredentialProvider
-func CreateProviderFromCredentials(ctx context.Context, registry string, credType CredType, creds map[string]string) common.RegistryCredentialProvider {
-	if len(creds) == 0 {
-		return common.NewPublicOnlyProvider()
-	}
-
-	// Create a callback provider that handles credentials based on type
-	callback := func(ctx context.Context, reg string, scope string) (*authn.AuthConfig, error) {
-		// Only handle the registry we're configured for
-		if reg != registry {
-			return nil, common.ErrNoCredentials
-		}
-
-		switch credType {
-		case CredTypeBasic:
-			// Basic auth with username/password
-			username := ""
-			password := ""
-
-			// Try different username keys
-			for key, value := range creds {
-				keyUpper := strings.ToUpper(key)
-				if strings.Contains(keyUpper, "USERNAME") {
-					username = value
-				}
-				if strings.Contains(keyUpper, "PASSWORD") {
-					password = value
-				}
-			}
-
-			if username != "" && password != "" {
-				return &authn.AuthConfig{
-					Username: username,
-					Password: password,
-				}, nil
-			}
-			return nil, common.ErrNoCredentials
-
-		case CredTypeAWS, CredTypeGCP, CredTypeAzure:
-			// For cloud providers, we need to set env vars temporarily for the keychain to work
-			// This is unavoidable as the AWS/GCP/Azure SDKs read from environment
-			for key, value := range creds {
-				os.Setenv(key, value)
-			}
-			keychain := common.NewKeychainProvider()
-			return keychain.GetCredentials(ctx, reg, scope)
-
-		case CredTypeToken:
-			// For token-based auth, use token as password with oauth2accesstoken username
-			for _, key := range tokenCredKeys {
-				if token, ok := creds[key]; ok && token != "" {
-					return &authn.AuthConfig{
-						Username: "oauth2accesstoken",
-						Password: token,
-					}, nil
-				}
-			}
-			return nil, common.ErrNoCredentials
-
-		default:
-			return nil, common.ErrNoCredentials
-		}
-	}
-
-	return common.NewCallbackProviderWithName(fmt.Sprintf("creds-%s", registry), callback)
-}
-
-// CreateProviderFromEnv creates a CLIP-compatible credential provider from environment variables
-// This function reads credentials from the environment based on the provided keys and creates
-// an appropriate credential provider for OCI registries
-// Returns common.RegistryCredentialProvider
-func CreateProviderFromEnv(ctx context.Context, registry string, credKeys []string) (common.RegistryCredentialProvider, error) {
-	if len(credKeys) == 0 {
-		return common.NewPublicOnlyProvider(), nil
-	}
-
-	// Build credential map from environment
-	creds := make(map[string]string)
-	for _, key := range credKeys {
-		if value := os.Getenv(key); value != "" {
-			creds[key] = value
-		}
-	}
-
-	if len(creds) == 0 {
-		return common.NewPublicOnlyProvider(), nil
-	}
-
-	// Detect credential type
-	credType := DetectCredentialType(registry, creds)
-
-	// Create a callback provider that handles credentials based on type
-	callback := func(ctx context.Context, reg string, scope string) (*authn.AuthConfig, error) {
-		// Only handle the registry we're configured for
-		if reg != registry {
-			return nil, common.ErrNoCredentials
-		}
-
-		switch credType {
-		case CredTypeBasic:
-			// Basic auth with username/password
-			username := ""
-			password := ""
-
-			// Try different username keys
-			for key, value := range creds {
-				keyUpper := strings.ToUpper(key)
-				if strings.Contains(keyUpper, "USERNAME") {
-					username = value
-				}
-				if strings.Contains(keyUpper, "PASSWORD") {
-					password = value
-				}
-			}
-
-			if username != "" && password != "" {
-				return &authn.AuthConfig{
-					Username: username,
-					Password: password,
-				}, nil
-			}
-			return nil, common.ErrNoCredentials
-
-		case CredTypeAWS:
-			// For AWS ECR, set AWS credentials and use keychain (handles ECR automatically)
-			if accessKey, ok := creds["AWS_ACCESS_KEY_ID"]; ok {
-				os.Setenv("AWS_ACCESS_KEY_ID", accessKey)
-			}
-			if secretKey, ok := creds["AWS_SECRET_ACCESS_KEY"]; ok {
-				os.Setenv("AWS_SECRET_ACCESS_KEY", secretKey)
-			}
-			if sessionToken, ok := creds["AWS_SESSION_TOKEN"]; ok {
-				os.Setenv("AWS_SESSION_TOKEN", sessionToken)
-			}
-			if region, ok := creds["AWS_REGION"]; ok {
-				os.Setenv("AWS_REGION", region)
-			}
-
-			// Use keychain provider which handles ECR
-			keychain := common.NewKeychainProvider()
-			return keychain.GetCredentials(ctx, reg, scope)
-
-		case CredTypeGCP:
-			// For GCP GCR, set GCP credentials and use keychain (handles GCR automatically)
-			if token, ok := creds["GCP_ACCESS_TOKEN"]; ok {
-				os.Setenv("GCP_ACCESS_TOKEN", token)
-			}
-			if credFile, ok := creds["GOOGLE_APPLICATION_CREDENTIALS"]; ok {
-				os.Setenv("GOOGLE_APPLICATION_CREDENTIALS", credFile)
-			}
-			if project, ok := creds["GCP_PROJECT_ID"]; ok {
-				os.Setenv("GCP_PROJECT_ID", project)
-			}
-
-			// Use keychain provider which handles GCR
-			keychain := common.NewKeychainProvider()
-			return keychain.GetCredentials(ctx, reg, scope)
-
-		case CredTypeAzure:
-			// For Azure ACR, set Azure credentials and use keychain
-			if clientID, ok := creds["AZURE_CLIENT_ID"]; ok {
-				os.Setenv("AZURE_CLIENT_ID", clientID)
-			}
-			if clientSecret, ok := creds["AZURE_CLIENT_SECRET"]; ok {
-				os.Setenv("AZURE_CLIENT_SECRET", clientSecret)
-			}
-			if tenantID, ok := creds["AZURE_TENANT_ID"]; ok {
-				os.Setenv("AZURE_TENANT_ID", tenantID)
-			}
-
-			// Use keychain provider which handles ACR
-			keychain := common.NewKeychainProvider()
-			return keychain.GetCredentials(ctx, reg, scope)
-
-		case CredTypeToken:
-			// For token-based auth, use token as password with oauth2accesstoken username
-			for _, key := range tokenCredKeys {
-				if token, ok := creds[key]; ok && token != "" {
-					return &authn.AuthConfig{
-						Username: "oauth2accesstoken",
-						Password: token,
-					}, nil
-				}
-			}
-			return nil, common.ErrNoCredentials
-
-		default:
-			return nil, common.ErrNoCredentials
-		}
-	}
-
-	return common.NewCallbackProviderWithName(fmt.Sprintf("env-%s", registry), callback), nil
-}
-
-// UnmarshalCredentials parses a JSON credential string
-func UnmarshalCredentials(credentialJSON string) (registry string, credType CredType, creds map[string]string, err error) {
-	var data map[string]interface{}
-
-	if err := json.Unmarshal([]byte(credentialJSON), &data); err != nil {
-		return "", "", nil, fmt.Errorf("failed to unmarshal credentials: %w", err)
-	}
-
-	registry, _ = data["registry"].(string)
-	credTypeStr, _ := data["type"].(string)
-	credType = CredType(credTypeStr)
-
-	creds = make(map[string]string)
-	if credMap, ok := data["credentials"].(map[string]interface{}); ok {
-		for k, v := range credMap {
-			if strVal, ok := v.(string); ok {
-				creds[k] = strVal
-			}
-		}
-	}
-
-	return registry, credType, creds, nil
-}
